addressed most of the suggested changes
