+++ src/components/Messages/ConversationList.tsx
@@ -9,7 +9,7 @@ import {
   setSelectedConversation,
 } from "../../state/slices/messagesSlice";
 import ChatOutlinedIcon from "@mui/icons-material/ChatOutlined";
-import SearchBar from "../Common/SearchBar";
+import SearchBar from "../Common/SearchBarDropdown";
 import CreateMessageModal from "./CreateMessageModal/CreateMessageModal";
 
 const styles = {
